requirements.txt
aiosignal==1.3.1
astropy==6.0.0
astropy-iers-data==0.2024.2.5.0.30.52
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
easyocr==1.7.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.2.0
h5py==3.10.0
idna==3.7
imageio==2.33.1
imbalanced-learn==0.12.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.3.0
protobuf==4.25.2
pyarrow==15.0.0
pyclipper==1.3.0.post5
pyerfa==2.0.1.1
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
ray==2.9.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
shap==0.44.1
shapely==2.0.2
six==1.16.0
slicer==0.0.7
sympy==1.12
threadpoolctl==3.2.0
tifffile==2024.1.30
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.4
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
xgboost==2.0.3
